server/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image Processing
# changed to >= to avoid "version not found" errors if 12.0.0 isn't available for your OS
Pillow>=10.2.0  
# You only need ONE OpenCV. We keep the non-headless one for the Camera Scanner GUI.
opencv-python==4.9.0.80

# Face Detection & Recognition
mediapipe==0.10.31
# dlib-bin is a pre-compiled version for Windows to avoid the C++ errors you saw earlier
dlib-bin; sys_platform == "win32"
face-recognition==1.3.0

# Object Detection (YOLO)
ultralytics==8.1.0

# Database
sqlalchemy==2.0.25

# Data Processing
numpy==1.26.3
pydantic==2.5.3

# Security
cryptography==41.0.7

# Testing
pytest==7.4.4
httpx==0.26.0
pillow-heif